+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to thsi project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added "license" and "changelog" commands, prints relative files
+
 ## [0.4.2] - 2018-05-02
 ### Added
 - Added "help" with search terms
@@ -42,7 +46,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.2.0] - 2017-04-27
 ### Added
-- Due day abbriviations supported
+- Due day abbreviations supported
 - Added archive command
 - Added report command